svg/text/parsing/text-anchor-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Text: parsing text-anchor with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/text.html#TextAnchorProperty" rel="help"/><h:meta content="text-anchor supports the full grammar 'start | middle | end'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("text-anchor", "start");
test_valid_value("text-anchor", "middle");
test_valid_value("text-anchor", "end");

  ]]></script></svg>
